refactor: standard 패키지를 global 패키지 내로 이동
