+++ src/main/java/com/ll/goohaeyou/domain/chat/message/controller/MessageController.java
@@ -6,7 +6,7 @@ import com.ll.goohaeyou.domain.chat.message.entity.Message;
 import com.ll.goohaeyou.domain.chat.message.service.MessageService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
 import com.ll.goohaeyou.global.security.MemberDetails;
-import com.ll.goohaeyou.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
