+++ src/main/java/com/ll/goohaeyou/domain/chat/room/controller/RoomController.java
@@ -5,7 +5,7 @@ import com.ll.goohaeyou.domain.chat.room.dto.RoomListDto;
 import com.ll.goohaeyou.domain.chat.room.service.RoomService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
 import com.ll.goohaeyou.global.security.MemberDetails;
-import com.ll.goohaeyou.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
