+++ src/main/java/com/ll/goohaeyou/domain/jobPost/employ/controller/EmployController.java
@@ -3,7 +3,7 @@ package com.ll.goohaeyou.domain.jobPost.employ.controller;
 import com.ll.goohaeyou.domain.application.application.dto.ApplicationDto;
 import com.ll.goohaeyou.domain.jobPost.employ.service.EmployService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
