+++ src/main/java/com/ll/goohaeyou/domain/jobPost/jobPost/controller/JobPostController.java
@@ -8,9 +8,9 @@ import com.ll.goohaeyou.domain.jobPost.jobPost.service.JobPostService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
 import com.ll.goohaeyou.global.config.AppConfig;
 import com.ll.goohaeyou.global.security.MemberDetails;
-import com.ll.goohaeyou.standard.base.Empty;
-import com.ll.goohaeyou.standard.base.util.CookieUtil;
-import com.ll.goohaeyou.standard.dto.PageDto;
+import com.ll.goohaeyou.global.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.util.CookieUtil;
+import com.ll.goohaeyou.global.standard.dto.PageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.Cookie;
