+++ src/main/java/com/ll/goohaeyou/domain/jobPost/jobPost/service/JobPostService.java
@@ -15,8 +15,8 @@ import com.ll.goohaeyou.domain.member.member.entity.type.Role;
 import com.ll.goohaeyou.domain.member.member.service.MemberService;
 import com.ll.goohaeyou.global.event.notification.*;
 import com.ll.goohaeyou.global.exception.CustomException;
-import com.ll.goohaeyou.standard.base.RegionType;
-import com.ll.goohaeyou.standard.base.util.Ut;
+import com.ll.goohaeyou.global.standard.base.RegionType;
+import com.ll.goohaeyou.global.standard.base.util.Ut;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.event.EventListener;
