+++ src/main/java/com/ll/goohaeyou/domain/member/emailAuth/controller/EmailAuthController.java
@@ -4,7 +4,7 @@ package com.ll.goohaeyou.domain.member.emailAuth.controller;
 import com.ll.goohaeyou.domain.member.emailAuth.service.EmailAuthService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
 import com.ll.goohaeyou.global.security.MemberDetails;
-import com.ll.goohaeyou.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
