+++ src/main/java/com/ll/goohaeyou/domain/member/member/controller/MemberController.java
@@ -9,7 +9,7 @@ import com.ll.goohaeyou.domain.member.member.service.MemberService;
 import com.ll.goohaeyou.domain.notification.service.NotificationService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
 import com.ll.goohaeyou.global.security.MemberDetails;
-import com.ll.goohaeyou.standard.base.util.CookieUtil;
+import com.ll.goohaeyou.global.standard.base.util.CookieUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
