+++ src/main/java/com/ll/goohaeyou/domain/payment/payment/service/PaymentService.java
@@ -12,8 +12,8 @@ import com.ll.goohaeyou.domain.payment.payment.entity.type.PayStatus;
 import com.ll.goohaeyou.domain.payment.payment.entity.repository.PaymentRepository;
 import com.ll.goohaeyou.global.config.TossPaymentsConfig;
 import com.ll.goohaeyou.global.exception.CustomException;
-import com.ll.goohaeyou.standard.base.util.TossPaymentUtil;
-import com.ll.goohaeyou.standard.retryOnOptimisticLock.RetryOnOptimisticLock;
+import com.ll.goohaeyou.global.standard.base.util.TossPaymentUtil;
+import com.ll.goohaeyou.global.standard.retryOnOptimisticLock.RetryOnOptimisticLock;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.minidev.json.JSONObject;
