+++ src/main/java/com/ll/goohaeyou/global/apiResponse/ApiResponse.java
@@ -3,7 +3,7 @@ package com.ll.goohaeyou.global.apiResponse;
 import com.ll.goohaeyou.global.exception.CustomHttpStatus;
 import com.ll.goohaeyou.global.exception.ErrorCode;
 import com.ll.goohaeyou.global.exception.type.ApiResultType;
-import com.ll.goohaeyou.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.Empty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
