+++ src/main/java/com/ll/goohaeyou/global/exception/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.ll.goohaeyou.global.exception;
 
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.standard.base.Empty;
+import com.ll.goohaeyou.global.standard.base.Empty;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
