+++ src/main/java/com/ll/goohaeyou/global/security/OAuth/OAuth2SuccessHandler.java
@@ -5,7 +5,7 @@ import com.ll.goohaeyou.domain.member.member.entity.RefreshToken;
 import com.ll.goohaeyou.domain.member.member.entity.repository.RefreshTokenRepository;
 import com.ll.goohaeyou.domain.member.member.service.MemberService;
 import com.ll.goohaeyou.global.security.JwtTokenProvider;
-import com.ll.goohaeyou.standard.base.util.CookieUtil;
+import com.ll.goohaeyou.global.standard.base.util.CookieUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
