+++ src/main/java/com/ll/goohaeyou/global/standard/base/Empty.java
@@ -1,0 +1,4 @@
+package com.ll.goohaeyou.global.standard.base;
+
+public class Empty {
+}
