+++ src/main/java/com/ll/goohaeyou/global/standard/base/KwType.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.standard.base;
+package com.ll.goohaeyou.global.standard.base;
 
 import lombok.AllArgsConstructor;
