+++ src/main/java/com/ll/goohaeyou/global/standard/base/util/MIMETypeUtil.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.standard.base.util;
+package com.ll.goohaeyou.global.standard.base.util;
 
 import java.util.HashMap;
 import java.util.Map;
