+++ src/main/java/com/ll/goohaeyou/global/standard/base/util/TossPaymentUtil.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.standard.base.util;
+package com.ll.goohaeyou.global.standard.base.util;
 
 import com.ll.goohaeyou.global.config.TossPaymentsConfig;
 import lombok.RequiredArgsConstructor;
