--- src/main/java/com/ll/goohaeyou/standard/base/Empty.java
@@ -1,4 +1,0 @@
-package com.ll.goohaeyou.standard.base;
-
-public class Empty {
-}
